third_party/WebKit/LayoutTests/fast/canvas/radialGradient-infinite-values-expected.txt

This test checks createRadialGradient with infinite values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createRadialGradient(0, 0, 100, 0, 0, NaN) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 6 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, 0, 0, Infinity) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 6 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, 0, 0, -Infinity) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 6 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, 0, NaN, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 5 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, 0, Infinity, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 5 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, 0, -Infinity, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 5 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, NaN, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 4 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, Infinity, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 4 is non-finite..
PASS ctx.createRadialGradient(0, 0, 100, -Infinity, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 4 is non-finite..
PASS ctx.createRadialGradient(0, 0, NaN, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 3 is non-finite..
PASS ctx.createRadialGradient(0, 0, Infinity, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 3 is non-finite..
PASS ctx.createRadialGradient(0, 0, -Infinity, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 3 is non-finite..
PASS ctx.createRadialGradient(0, NaN, 100, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 2 is non-finite..
PASS ctx.createRadialGradient(0, Infinity, 100, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 2 is non-finite..
PASS ctx.createRadialGradient(0, -Infinity, 100, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 2 is non-finite..
PASS ctx.createRadialGradient(NaN, 0, 100, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 1 is non-finite..
PASS ctx.createRadialGradient(Infinity, 0, 100, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 1 is non-finite..
PASS ctx.createRadialGradient(-Infinity, 0, 100, 0, 0, 100) threw exception TypeError: Failed to execute 'createRadialGradient' on 'CanvasRenderingContext2D': float parameter 1 is non-finite..
PASS successfullyParsed is true

TEST COMPLETE
